requirements.txt
pandas
numpy
scikit-learn
mlflow
fastapi
uvicorn[standard]
pydantic
pyyaml
joblib
matplotlib
seaborn